Reservation-MicroService/src/main/java/com/reservation/service/global/GlobalExceptionHandler.java
package com.reservation.service.global;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ExceptionHandler;

import com.reservation.service.exception.InvalidReservationIdException;

public class GlobalExceptionHandler {
	@ExceptionHandler(Exception.class)
	public final ResponseEntity<?> handleAllExceptions(Exception ex) {

		return new ResponseEntity<>(ex.getMessage(), HttpStatus.UNAUTHORIZED);
	}
	@ExceptionHandler(InvalidReservationIdException.class)
	public final ResponseEntity<?> handleNotFoundException(InvalidReservationIdException ex) {
		return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
	}
	
}